Moved Fusion Solar related stuff into separate folder
