+++ custom_components/fusion_solar_kiosk/const.py
@@ -6,18 +6,6 @@ DOMAIN = 'fusion_solar_kiosk'
 CONF_KIOSKS = 'kiosks'
 CONF_KIOSK_URL = 'url'
 
-# Fusion Solar Kiosk API response attributes
-ATTR_DATA = 'data'
-ATTR_FAIL_CODE = 'failCode'
-ATTR_SUCCESS = 'success'
-ATTR_DATA_REALKPI = 'realKpi'
-# Data attributes
-ATTR_REALTIME_POWER = 'realTimePower'
-ATTR_TOTAL_CURRENT_DAY_ENERGY = 'dailyEnergy'
-ATTR_TOTAL_CURRENT_MONTH_ENERGY = 'monthEnergy'
-ATTR_TOTAL_CURRENT_YEAR_ENERGY = 'yearEnergy'
-ATTR_TOTAL_LIFETIME_ENERGY = 'cumulativeEnergy'
-
 # Possible ID suffixes
 ID_REALTIME_POWER = 'realtime_power'
 ID_TOTAL_CURRENT_DAY_ENERGY = 'total_current_day_energy'
