+++ custom_components/fusion_solar_kiosk/fusion_solar/const.py
@@ -1,0 +1,12 @@
+# Fusion Solar Kiosk API response attributes
+ATTR_DATA = 'data'
+ATTR_DATA_REALKPI = 'realKpi'
+ATTR_FAIL_CODE = 'failCode'
+ATTR_SUCCESS = 'success'
+
+# Data attributes
+ATTR_REALTIME_POWER = 'realTimePower'
+ATTR_TOTAL_CURRENT_DAY_ENERGY = 'dailyEnergy'
+ATTR_TOTAL_CURRENT_MONTH_ENERGY = 'monthEnergy'
+ATTR_TOTAL_CURRENT_YEAR_ENERGY = 'yearEnergy'
+ATTR_TOTAL_LIFETIME_ENERGY = 'cumulativeEnergy'
