+++ custom_components/fusion_solar_kiosk/sensor.py
@@ -3,19 +3,26 @@ import homeassistant.helpers.config_validation as cv
 import logging
 import voluptuous as vol
 
-from .fusion_solar.kiosk_api import FusionSolarKioksApi
-from . import FusionSolarKioskEnergyEntity, FusionSolarKioskPowerEntity
-
 from datetime import timedelta
-from homeassistant.components.sensor import PLATFORM_SCHEMA
-from homeassistant.const import (
-    CONF_NAME,
-)
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
-from .const import *
 
-import re
-from urllib.parse import urlparse
+from homeassistant.components.sensor import PLATFORM_SCHEMA
+from homeassistant.const import CONF_NAME
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
+
+from .fusion_solar.const import ATTR_DATA_REALKPI, ATTR_REALTIME_POWER, ATTR_TOTAL_CURRENT_DAY_ENERGY, \
+    ATTR_TOTAL_CURRENT_MONTH_ENERGY, ATTR_TOTAL_CURRENT_YEAR_ENERGY, ATTR_TOTAL_LIFETIME_ENERGY
+from .fusion_solar.kiosk import Kiosk
+from .fusion_solar.kiosk_api import FusionSolarKioksApi
+from .fusion_solar.energy_entity import FusionSolarKioskSensorTotalCurrentDayEnergy, \
+    FusionSolarKioskSensorTotalCurrentMonthEnergy, FusionSolarKioskSensorTotalCurrentYearEnergy, \
+    FusionSolarKioskSensorTotalLifetimeEnergy
+from .fusion_solar.power_entity import FusionSolarKioskSensorRealtimePower
+
+from .const import CONF_KIOSK_URL, CONF_KIOSKS, ID_REALTIME_POWER, NAME_REALTIME_POWER, \
+    ID_TOTAL_CURRENT_DAY_ENERGY, NAME_TOTAL_CURRENT_DAY_ENERGY, \
+    ID_TOTAL_CURRENT_MONTH_ENERGY, NAME_TOTAL_CURRENT_MONTH_ENERGY, \
+    ID_TOTAL_CURRENT_YEAR_ENERGY, NAME_TOTAL_CURRENT_YEAR_ENERGY, \
+    ID_TOTAL_LIFETIME_ENERGY, NAME_TOTAL_LIFETIME_ENERGY
 
 KIOSK_SCHEMA = vol.Schema(
     {
@@ -101,41 +108,3 @@ async def async_setup_platform(hass, config, async_add_entities, discovery_info=
                 ATTR_TOTAL_LIFETIME_ENERGY,
             )
         ])
-
-
-class Kiosk:
-    def __init__(self, url, name):
-        self.url = url
-        self.name = name
-        self._parseId()
-
-    def _parseId(self):
-        id = re.search("\?kk=(.*)", self.url).group(1)
-        _LOGGER.debug('calculated KioskId: ' + id)
-        self.id = id
-
-    def apiUrl(self):
-        url = urlparse(self.url)
-        apiUrl = (url.scheme + "://" + url.netloc)
-        _LOGGER.debug('calculated API base url for ' + self.id + ': ' + apiUrl)
-        return apiUrl
-
-
-class FusionSolarKioskSensorRealtimePower(FusionSolarKioskPowerEntity):
-    pass
-
-
-class FusionSolarKioskSensorTotalCurrentDayEnergy(FusionSolarKioskEnergyEntity):
-    pass
-
-
-class FusionSolarKioskSensorTotalCurrentMonthEnergy(FusionSolarKioskEnergyEntity):
-    pass
-
-
-class FusionSolarKioskSensorTotalCurrentYearEnergy(FusionSolarKioskEnergyEntity):
-    pass
-
-
-class FusionSolarKioskSensorTotalLifetimeEnergy(FusionSolarKioskEnergyEntity):
-    pass
